requirements.txt
argcomplete==3.2.2
certifi==2024.2.2
charset-normalizer==3.3.2
idna==3.6
ijson==3.2.3
markdown-it-py==3.0.0
mdurl==0.1.2
pyfiglet==1.0.2
Pygments==2.17.2
python-decouple==3.8
requests==2.31.0
rich==13.7.1
urllib3==2.2.1
